[heap] Fix AllocationLimitOvershotByLargeMargin underflow
The assumption that allocation_limit < max_size isn't always true.
This won't hold if current size > max_size.
When this happens for old generation limit, we'll eventually crash, but that's not the case for global limit.
As a result, part of global_overshoot_margin() can underflow and incorrectly return `limit / 2` as a fallback (and AllocationLimitOvershotByLargeMargin would incorrectly return false).
This CL avoids the underflow by forcing AllocationLimitOvershotByLargeMargin() = true as soon as allocation_limit > max_size.
In practise, this doesn't change a whole lot because CanExpandOldGeneration() would also cause a slow allocation to fail.

Bug: 361124432
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7500073
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#105587}

+++ src/heap/heap.h
@@ -1444,7 +1444,7 @@ class Heap final {
   V8_EXPORT_PRIVATE size_t EmbedderSizeOfObjects() const;
 
   // Returns the global size of objects (embedder + V8 non-new spaces).
-  V8_EXPORT_PRIVATE size_t GlobalSizeOfObjects() const;
+  V8_EXPORT_PRIVATE uint64_t GlobalSizeOfObjects() const;
 
   // Returns the global amount of wasted bytes.
   V8_EXPORT_PRIVATE size_t GlobalWastedBytes() const;
